+++ src/macros/mod.rs
@@ -5,9 +5,13 @@ mod fs;
 mod io;
 mod string;
 mod system;
+mod text;
 
 #[cfg(feature = "serde")]
 mod json;
 
 #[cfg(feature = "time")]
 mod time;
+
+#[cfg(feature = "async")]
+mod r#async;